+++ packages/backend/src/storage/storage.ts
@@ -45,6 +45,7 @@ import { IpfsFilesManager, IpfsFilesManagerEvents } from './ipfsFileManager'
 import { create } from 'ipfs-core'
 
 import { CID } from 'multiformats/cid'
+import { NoCryptoEngineError } from '@quiet/types'
 
 const log = logger('db')
 
@@ -257,7 +258,7 @@ export class Storage extends EventEmitter {
     log('createDbForChannels init')
     this.channels = await this.orbitdb.keyvalue<PublicChannel>('public-channels', {
       accessController: {
-        type: 'channelsaccess',
+        // type: 'channelsaccess',
         write: ['*']
       }
     })
@@ -276,13 +277,13 @@ export class Storage extends EventEmitter {
       })
 
       // Delete channel on replication
-      Array.from(this.publicChannelsRepos.keys()).forEach(e => {
-        const isDeleted = !Object.keys(this.channels.all).includes(e as string)
-        if (isDeleted) {
-          log('deleting channel ', e)
-          void this.deleteChannel({ channel: e })
-        }
-      })
+      // Array.from(this.publicChannelsRepos.keys()).forEach(e => {
+      //   const isDeleted = !Object.keys(this.channels.all).includes(e as string)
+      //   if (isDeleted) {
+      //     log('deleting channel ', e)
+      //     void this.deleteChannel({ channel: e })
+      //   }
+      // })
 
       Object.values(this.channels.all).forEach(async (channel: PublicChannel) => {
         await this.subscribeToChannel(channel)
@@ -340,6 +341,8 @@ export class Storage extends EventEmitter {
 
   async verifyMessage(message: ChannelMessage): Promise<boolean> {
     const crypto = getCrypto()
+    if (!crypto) throw new NoCryptoEngineError()
+
     const signature = stringToArrayBuffer(message.signature)
     let cryptoKey = this.publicKeysMap.get(message.pubKey)
 
@@ -509,7 +512,22 @@ export class Storage extends EventEmitter {
     if (channel) {
       void this.channels.del(payload.channel)
     }
-    const repo = this.publicChannelsRepos.get(payload.channel)
+    let repo = this.publicChannelsRepos.get(payload.channel)
+    if (!repo) {
+      const db = await this.orbitdb.log<ChannelMessage>(
+        `channels.${payload.channel}`,
+        {
+          accessController: {
+            type: 'messagesaccess',
+            write: ['*']
+          }
+        }
+      )
+      repo = {
+        db,
+        eventsAttached: false
+      }
+    }
     await repo.db.load()
     const allEntries = this.getAllEventLogRawEntries(repo.db)
     await repo.db.close()
@@ -523,7 +541,7 @@ export class Storage extends EventEmitter {
     await this.deleteChannelFiles(files)
     await this.deleteChannelMessages(hashes)
     this.publicChannelsRepos.delete(payload.channel)
-    this.emit(StorageEvents.DELETED_CHANNEL, payload)
+    this.emit(StorageEvents.CHANNEL_DELETION_RESPONSE, payload)
   }
 
   public async deleteChannelFiles(files: FileMetadata[]) {