+++ packages/mobile/src/components/ChannelList/ChannelList.stories.tsx
@@ -17,9 +17,7 @@ storiesOf('ChannelList', module).add('Default', () => (
           'Text from latest chat message. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Id massa venenatis id eget massa commodo posuere faucibus aliquam. At scelerisque nisi mauris facilisis.',
         date: '1:55pm',
         unread: false,
-        redirect: (address: string) => { console.log(`Clicked ${address}`) },
-        deleteChannel: (channel: string) => { console.log(`Deleted channel ${channel}`) },
-        enableDeletion: true
+        redirect: (address: string) => { console.log(`Clicked ${address}`) }
       },
       {
         name: 'spam',
@@ -28,9 +26,7 @@ storiesOf('ChannelList', module).add('Default', () => (
           'Text from latest chat message. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Id massa venenatis id eget massa commodo posuere faucibus aliquam. At scelerisque nisi mauris facilisis.',
         date: '1:55pm',
         unread: false,
-        redirect: (address: string) => { console.log(`Clicked ${address}`) },
-        deleteChannel: (channel: string) => { console.log(`Deleted channel ${channel}`) },
-        enableDeletion: true
+        redirect: (address: string) => { console.log(`Clicked ${address}`) }
       },
       {
         name: 'design',
@@ -39,9 +35,7 @@ storiesOf('ChannelList', module).add('Default', () => (
           'Text from latest chat message. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Id massa venenatis id eget massa commodo posuere faucibus aliquam. At scelerisque nisi mauris facilisis.',
         date: '6/1/22',
         unread: true,
-        redirect: (address: string) => { console.log(`Clicked ${address}`) },
-        deleteChannel: (channel: string) => { console.log(`Deleted channel ${channel}`) },
-        enableDeletion: true
+        redirect: (address: string) => { console.log(`Clicked ${address}`) }
       },
       {
         name: 'qa',
@@ -50,9 +44,7 @@ storiesOf('ChannelList', module).add('Default', () => (
           'Text from latest chat message. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Id massa venenatis id eget massa commodo posuere faucibus aliquam. At scelerisque nisi mauris facilisis.',
         date: 'Yesterday',
         unread: false,
-        redirect: (address: string) => { console.log(`Clicked ${address}`) },
-        deleteChannel: (channel: string) => { console.log(`Deleted channel ${channel}`) },
-        enableDeletion: true
+        redirect: (address: string) => { console.log(`Clicked ${address}`) }
       }
     ]}
   />