+++ packages/mobile/src/components/ChannelTile/ChannelTile.stories.tsx
@@ -14,21 +14,6 @@ storiesOf('ChannelTile', module)
       date={'1:55pm'}
       unread={false}
       redirect={(address: string) => { console.log(`Clicked ${address}`) }}
-      deleteChannel={(channel: string) => { console.log(`Deleted channel ${channel}`) }}
-    />
-  ))
-  .add('Deletion', () => (
-    <ChannelTile
-      name={'general'}
-      address={'general'}
-      message={
-        'Text from latest chat message. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Id massa venenatis id eget massa commodo posuere faucibus aliquam. At scelerisque nisi mauris facilisis.'
-      }
-      date={'1:55pm'}
-      unread={false}
-      redirect={(address: string) => { console.log(`Clicked ${address}`) }}
-      deleteChannel={(channel: string) => { console.log(`Deleted channel ${channel}`) }}
-      enableDeletion
     />
   ))
   .add('Unread', () => (
@@ -41,6 +26,5 @@ storiesOf('ChannelTile', module)
       date={'1:55pm'}
       unread={true}
       redirect={(address: string) => { console.log(`Clicked ${address}`) }}
-      deleteChannel={(channel: string) => { console.log(`Deleted channel ${channel}`) }}
     />
   ))