+++ packages/state-manager/src/sagas/publicChannels/channelDeletionResponse/channelDeletionResponse.saga.ts
@@ -3,13 +3,12 @@ import { PayloadAction } from '@reduxjs/toolkit'
 import logger from '../../../utils/logger'
 import { put, delay, select } from 'typed-redux-saga'
 import { messagesActions } from '../../messages/messages.slice'
-import { MessageType, WriteMessagePayload } from '../../messages/messages.types'
 import { communitiesSelectors } from '../../communities/communities.selectors'
 
 const log = logger('publicChannels')
 
-export function* deletedChannelSaga(
-  action: PayloadAction<ReturnType<typeof publicChannelsActions.deletedChannel>['payload']>
+export function* channelDeletionResponseSaga(
+  action: PayloadAction<ReturnType<typeof publicChannelsActions.channelDeletionResponse>['payload']>
 ): Generator {
   log(`Deleted channel ${action.payload.channel} saga`)
 
@@ -28,33 +27,16 @@ export function* deletedChannelSaga(
 
   yield* put(publicChannelsActions.deleteChannelFromStore({ channelAddress }))
 
-  const ownerNickname = yield* select(communitiesSelectors.ownerNickname)
-
   const community = yield* select(communitiesSelectors.currentCommunity)
 
   const isOwner = Boolean(community?.CA)
 
-  let message: string
-
-  if (isGeneral) {
-    if (isOwner) {
-      yield* put(publicChannelsActions.createGeneralChannel())
-    }
-    // For better UX
-    yield* delay(500)
-    yield* put(publicChannelsActions.finishGeneralRecreation())
-    message = `#general has been recreated by @${ownerNickname}`
-  } else {
-    message = `@${ownerNickname} deleted #${channelAddress}`
-  }
-
-  const payload: WriteMessagePayload = {
-    type: MessageType.Info,
-    message,
-    channelAddress: 'general'
-  }
-
   if (isOwner) {
-    yield* put(messagesActions.sendMessage(payload))
+    if (isGeneral) {
+      yield* delay(1000)
+      yield* put(publicChannelsActions.createGeneralChannel())
+    } else {
+      yield* put(messagesActions.sendDeletionMessage({ channelAddress }))
+    }
   }
 }