+++ packages/common/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1-alpha.2](https://github.com/TryQuiet/quiet/compare/@quiet/common@1.2.1-alpha.1...@quiet/common@1.2.1-alpha.2) (2023-04-28)
+
+**Note:** Version bump only for package @quiet/common
+
+
+
+
+
+## [1.2.1-alpha.1](https://github.com/TryQuiet/quiet/compare/@quiet/common@1.2.1-alpha.0...@quiet/common@1.2.1-alpha.1) (2023-04-27)
+
+**Note:** Version bump only for package @quiet/common
+
+
+
+
+
 ## [1.2.1-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/common@1.2.0...@quiet/common@1.2.1-alpha.0) (2023-04-26)
 
 **Note:** Version bump only for package @quiet/common